requirements.txt
python-telegram-bot
pandas
matplotlib
seaborn
openpyxl
xlrd